backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx>=0.27,<1
pytest==7.4.4
pytest-asyncio==0.23.6

# Database integrations
cassandra-driver==3.28.0
aiohttp==3.9.5

# Vector embeddings and ML (Phase 2)
sentence-transformers==2.7.0
numpy==1.26.4
scikit-learn==1.4.2
openai
aiofiles==23.2.1

# LLM Integration (Phase 3)
ollama==0.5.1
anthropic==0.54.0
litellm==1.72.6
tiktoken==0.9.0
sentencepiece==0.1.99

# SREC dependencies
Wand==0.6.13
watchdog==3.0.0

# Kafka for future
# kafka-python==2.0.2